+++ Support_Your_Locals/Infrastructure/Extensions/UrlExtensions.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Http;
 
-namespace Support_Your_Locals.Infrastructure
+namespace Support_Your_Locals.Infrastructure.Extensions
 {
     public static class UrlExtensions
     {
+
         public static string PathAndQuery(this HttpRequest request) =>
             request.QueryString.HasValue ? $"{request.Path}{request.QueryString}" : request.Path.ToString();
     }
