+++ Support_Your_Locals/Views/_ViewImports.cshtml
@@ -2,6 +2,7 @@
 @using Support_Your_Locals.Models.ViewModels
 @using Support_Your_Locals.Models.ViewModels.BusinessBoard
 @using Support_Your_Locals.Infrastructure
+@using Support_Your_Locals.Infrastructure.Extensions
 @using Support_Your_Locals.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Support_Your_Locals
